pyproject.toml
[project]
name = "finance-tracker-app"
version = "0.1.0"
description = "Finance Tracker CLI integrated with a google sheet"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
ft = "financetracker.main:cli"
